XCF-1632 add Opera specific CSS stylesheet (for woof!)
Add opera-hacks.css that is only loaded when Opera browser is used.
Opera does not render <fieldset/> in the same way that other browsers.
Add 'display: table' property to Opera specific CSS file so that Opera
does not leave an empty space at the top of the page whenever we have
a floating help box.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10643 ab1d8caa-1f67-47f1-9e81-24633a41865c