XCF-2236: Fixed the generated authrules.xml.in file so that it contains the correct <userPattern> elements.
When DialingRules are created on the fly for the emergency routing, the default CallPattern constructor was used, setting the CallDigits to NO_DIGITS, causing an empty pattern string to be returned.

I added an explicit call to CallPattern(String, CallDigits) to set the CallDigits to FIXED_DIGITS, which resulted in the correct pattern (the external emergency number) being returned.

Also added a unit test method to AuthRulesTest to ensure that the generated XML contains the <userPattern>.

Patch submitted by Joe Attardi.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11972 ab1d8caa-1f67-47f1-9e81-24633a41865c