XCF-2028 add logging and SIP logging level setting to sipXpage
New advanced section on Paging Groups:
contains a "SIP Trace Level" combo box with default value NONE. This value is save in
sipxpage.properties.in when the "Activate" button is pressed.

The "SIP Trace Level" value is saved in paging_server table when the "Save" button is pressed and it
is replicated in sipxpage.properties.in file when configuration is activated.

Patch from Laurentiu Ceausescu - thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12735 ab1d8caa-1f67-47f1-9e81-24633a41865c