XCF-2621: Added conference control to the user portal.
The administrator can see the conferences assigned to a user under the "Conferences" tab when editing a user.
When a user is logged in, "My Information" --> "Conferences" lists the user's assigned conferences, and the user can:
 - Fully control the conferences (lock/unlock, mute, deaf, kick, etc.)
 - Edit some conference properties (all except name and extension, which are admin editable only).

These changes are a prelude to XCF-828, which will allow conferences to be automatically assigned to new users.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12982 ab1d8caa-1f67-47f1-9e81-24633a41865c