Added org.sipfoundry.commons.siprouter to sipXcommons
Includes 
   ProxyRouter.java
   FindServer.java
   FindServerTest.java

These add psuedo RFC-3263 Locating SIP Servers support to JAIN-SIP users.

JAIN-SIP has to be told to use the ProxyRouter via this property:

   properties.setProperty("javax.sip.ROUTER_PATH", 
      org.sipfoundry.commons.siprouter.ProxyRouter.class.getName());

The algorithm is close to RFC-3263, but not quite.  It does use NAPTR,
SRV, and A records to find SIP servers, and picks the best transport
based on the RFC-3263 rules.  It does not do the weight or priority
parts, and doesn't keep track of failed requests.  It relies on DNSJava
for DNS lookup, caching, and Round Robin services.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13661 ab1d8caa-1f67-47f1-9e81-24633a41865c