XCF-3155: Fixed the descriptive tooltips for the configuration diagnostic tests. The problem turned out to be caused by Tapestry, and not Dojo. For some reason Tapestry was appending "_0" to the tooltip element IDs, which caused them to not be found by the calls to dojo.byId().
Furthermore, the automatic refresh now gets temporarily disabled while a tooltip is showing,
regardless of the state of the checkbox. This is to ensure the user gets enough time to
read the tooltip before the page refreshes and the tooltip disappears.

Lastly, it looks like the auto refresh checkbox wasn't even working. Even when unchecked,
the auto refresh was still happening. I have updated the AutoRefreshForm script to check the
state of this checkbox before refreshing.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14422 ab1d8caa-1f67-47f1-9e81-24633a41865c