XCF-3233 unescape Java string before passing into LdapName
The problem turns out to be passing a Java-escaped string into LdapName.
The apostrophe in the Distinguished Name is already escaped in
accordance with RFC 2253, but ends up with two backslashes (e.g.
O\\'Reilly) after Java-escaping, and this causes an
InvalidNameException. We have to unescapeJava before passing into
LdapName.

Patch from: Carolyn Beeton - thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14939 ab1d8caa-1f67-47f1-9e81-24633a41865c