XCF-3195 double nature of sipXbridge configuration
Since sipXbridge is both a device and a service in some contexts we need
to replicate it as a service (during start-up, when bundle is added) and
in other cases we need to replicate it as a device (when user sends
device profiles).

BridgeSbcConfigurationFile is an adapter that implements
ConfigurationFile interface but uses device infrastructure to replicate
the SBC config.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@15129 ab1d8caa-1f67-47f1-9e81-24633a41865c