XX-5377 new site-to-site rule
Can be used to connect 2 or more sipX installations.
Does not allow for any permissions and does not add any values to
authrules.xml.

Site-to-site rules only work if gateways (other sites) are *not* used in
rules of any other kind.

new IDialRule.isAuthorizationChecked - it is used to determine if the
rule should be used when generating authrules.xml
For now it's only false for 'site-to-site' rules.

Minor fix for EnumEditPageProfider - did not associate site-to-site
rules properly.

Patch from: Huijun Yang - thanks!

Merged from: https://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@15334

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@15396 ab1d8caa-1f67-47f1-9e81-24633a41865c