XX-5541 Add Click-to-call to the voicemail table on the user portal
From address becomes a link: clicking a link will initiate the call.
We are trying to parse address spec from the "From" portion of voicemail
specification. It seems to work with legacy voicemail but will have to
be retested after voicemail is upgraded.

Based on the patch from: Huijun Yang - thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@15809 ab1d8caa-1f67-47f1-9e81-24633a41865c