XX-7586: "SIPRegistrar" shows "Resources Required " on a fresh ISO installation of the build 4.1.6-017883
This was caused by 983ea213e9969fb6fad8c34ba1f0ae43115228ad.

Revert "XX-5994 use IMDB table to store presence routing preferences"

This reverts commit 983ea213e9969fb6fad8c34ba1f0ae43115228ad.

It was believed that registrar need to be restarted on every configuration changes to pick up the config changes. Other than IMDB, there is another option is to have a backgroud process to monitor configuration changes, and re-parse the configurations if it is changed. The latter option appears to be less effort than IMDB option in server side, thus revert back the IMDB changes in sipXconfig.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17887 ab1d8caa-1f67-47f1-9e81-24633a41865c