XX-6702  write active greeting (mailbox preference)to mailboxprefs.xml
SipXconfig will write in this file the mailbox preferences (as of this issue
just the active greeting) for sipXivr (and possible other clients) to read;
mailboxprefs.xml is "per user" file kept in the mailstore directory.

On sipXconfig side the active greeting is still kept in the DB as a user setting.

The file will be replicated when the active greeting is changed from admin and user portal and via the RESTful API.

The migration task will migrate mailboxprefs.xml from the old to the new format on upgrade
(imap stuff has been moved to validusers.xml as of 4.2)


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17996 ab1d8caa-1f67-47f1-9e81-24633a41865c