XX-7253 New sipXconfig User and Admin portal layout
Omnibus revision:
XX-7744 - The images used to create the rounded look needed to be wider to support larger monitors and resolutions.
XX-7937 - A margin spacing was removed to eliminate the white space (band) that was appearing on the component-submit elements.
XX-7632 - fixes the elements that were floating off to the right-hand side beyond the center container outline.
XX-7980 - lengthier content will now extend that area
XX-7789 - Modified each page with a tableDiv to apply the overflow style to the form.  With this, all forms inside that middle content area will add a scrollbar when it's required.

Patch from Mike Da Silva, thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@18413 ab1d8caa-1f67-47f1-9e81-24633a41865c