XX-7863 Scheduled Backup does not work if the services are restarted.
Problem was that Hibernate was being called directly so additional Spring parameters in the bean definition were not being sent when the object was created. This had the effect of causing the backup to fail completely and get stuck.

Resolved the issue by adding the Hibernate session factory property to the bean definition for AdminContextImpl. Calling Hibernate now will force the call through Spring so that all properties in the AdminContextImpl are set.

Patch from Raymond Dans, thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@18537 ab1d8caa-1f67-47f1-9e81-24633a41865c