XX-8303 - reset-superadmin does not work
If the 'sipxconfig.sh' script is running as 'root' user and 'mailstore' directory is missing, the 'mailstore' directory will be created with 'root' permissions.
So I modified 'sipxconfig.sh' script in order to change the owner of the 'mailstore' directory, after the sipXconfig code is called (which make the 'mailstore' directory).

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.2@18772 ab1d8caa-1f67-47f1-9e81-24633a41865c